pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "canvas-companion"
version = "0.1.0"
description = "Monitor Canvas LMS, sync files to Google Drive, send Telegram alerts"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "google-api-python-client>=2.100",
    "google-auth>=2.25",
    "google-auth-oauthlib>=1.2",
    "google-auth-httplib2>=0.2",
    "python-telegram-bot>=21.0",
    "APScheduler>=3.10,<4",
    "typer>=0.12",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "google-genai>=1.0",
    "pymupdf>=1.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "respx>=0.21",
    "ruff>=0.3",
]

[project.scripts]
canvas-companion = "canvas_companion.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/canvas_companion"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100
